+++ packages/core/src/internal/index.ts
@@ -1,12 +1,11 @@
 export type {
-  AbiParameter,
-  ContractConfig,
   ContractsConfig,
   ContractsResult,
   Event,
   GetArgs,
   GetConfig,
   GetReturnType,
+  Contract,
 } from '../types/contracts'
 export type { IsNever, NotEqual, Optional, Or } from '../types/utils'
